Initial commit.
This is **not a perfect Docker setup** by any means - I'm just trying to get to a point where I can start to building out the ColdFusion app. One particularly problematic note is that I am going to be running the JavaScript and LESS CSS compilers **outside of the container**; which, in many ways, defeats the whole point of containerization as my host machine now needs to have `npm` and `node` installed. But, containerization is not the goal of my exploration - I'm here to look at Hotwire.