Updating the healthcheck to point to a non-index page.
I was noticing that the `index.cfm` page was being hit every 20-seconds. Normally, this shouldn't matter; but, the index page will likely have queries and other logic that I don't want being hit all the time. As such, I am moving the healcheck to a dedicated template. Eventually, this will likely be moved to an event-based route; but, keeping it simple for now.