Move template-related values above form processing.
In order to allow me to `exit` out of the action in the delete-tippee template, I had to move the `request.template.title` above the form processing (otherwise it wouldn't be set in some cases). To keep things consistent across the app, I've moved it up in other actions as well.