Add and Edit workflows for Events.
These views deal with explicit dates, which is always problematic in web applications because they are submitted as Strings but need to be parsed as dates. I don't love that I am using an `isDate()` call outside of my normal validation; but, it's one of those tough things where ColdFusion will actually enforce data-types in the method calls, so I can't just _hope_ that the given value is date-like, I need to make sure it is before making the method call.