Add the ability to duplicate an event.
For the first pass of this, I'm going to automatically copy all tips over to the new event. However, in the future, it might be a nice user experience to see the list of tips - with amounts - that the user can change, or remove, prior to duplication. But, I don't want to go down that rabbit hole at this time.