Add tip list view.
For now, I'm not using any pagination - this just lists out all the tips in the system. Obviously, not a long-term strategy; but, enough to get started with. I've also added some Completed/Not-Completed calls-to-action in the list itself. I may end up removing those since the Dashboard will eventually act as the "command center" for pending tips.