(Partially) Add the workflow for creating a new Tip.
The user interface for this is super hacky, and incomplete. This version requires a new Tippee / Event to be created as part of the tip creation process. I'll update this to pull from records next. That said, even once I update it, the UI will still be super janky - but, I'm trying to get this all done without JavaScript.