Cross-link detail pages with tip-creation form.
This allows the user to jump directly from the Tippee/Event detail views into the tip creation page; with, an option to Cancel back to the detail page. I am opting not to return the user to detail page _after_ tip creation since my hunch is that people will want to see the list of tips after a new tip is created.