Installing Hotwire and setting up an minimal build.
For the sake of simplicity, I'm not hashing the filename or anything when the JavaScript files are generated. This way, I don't have to update the script tag - I just rely on my Chrome dev tools to disable cache during development. This isn't an exploration in proper file generation - it's an exploration of Hotwire.