Switch from `.cfm` extension to `.htm` extension.
After all of the CFML that I've been writing, it was a bit soul-crushing to finally install Hotwire only to find that it doesn't work with CFML. Or, rather, it doesn't work with `.cfm` extensions. This is by design, though there are talks on the Hotwire side to re-think this constraint:

**GitHub Issue**: [Remove isHTML](https://github.com/hotwired/turbo/pull/519)

Since non-HTML extensions might serve up _anything_, Hotwire can't know that said URL is for an HTML page. And, therefore, can't know that it has to intercept the link-click/form-submission. As such, I have to change the `.cfm` extensions to `.htm` extensions and then enable SES URL rewriting in the CommandBox server. This will rewrite to `index.cfm` with the `cgi.path_info` value being set to the original URL. That said, since I'm only using the `index.cfm` in this app, I don't even have to worry about it.