Turbo Drive requires non-2xx status on Form failures.
In order for Turbo Drive to interoperate properly with form submissions, a failed form must respond with a non-`2xx` status code. This is because Turbo Drive is expecting to **redirect the page** upon a successful form submission (which I think is part of the HTTP specification or suggested practice).

There are some bugs in this app with the complex forms; but, I'm hoping to fix those with Stimulus and Frames when I get there.